requirements.txt
pandas>=2.0
Pillow>=10.0
arabic-reshaper>=2.1.0
python-bidi>=0.4.2
gTTS>=2.3.2
moviepy>=2.0.0
googletrans==4.0.0-rc1
pydub>=0.25.1
python-dotenv>=1.0.0
tqdm>=4.65
